Finished implementation and cleaned up all compiler warnings
