+++ src/exp/list_exp.rs
@@ -1,14 +1,40 @@
 //! List Cdt Aerospike Filter Expressions.
 //!
-use crate::exp::exp::{ExpOp, ExpType, Expression, ExpressionArgument, FilterExpression, MODIFY};
+use crate::exp::{ExpOp, ExpType, Expression, ExpressionArgument, FilterExpression, MODIFY};
 use crate::operations::cdt_context::CdtContext;
 use crate::operations::lists::{ListPolicy, ListReturnType, ListSortFlags};
 use crate::Value;
 
+/// List expression generator.
+///
+/// The bin expression argument in these methods can be a reference to a bin or the
+/// result of another expression. Expressions that modify bin values are only used
+/// for temporary expression evaluation and are not permanently applied to the bin.
+///
+/// List modify expressions return the bin's value. This value will be a list except
+/// when the list is nested within a map. In that case, a map is returned for the
+/// list modify expression.
+///
+/// List expressions support negative indexing. If the index is negative, the
+/// resolved index starts backwards from end of list. If an index is out of bounds,
+/// a parameter error will be returned. If a range is partially out of bounds, the
+/// valid part of the range will be returned. Index/Range examples:
+///
+/// Index 0: First item in list.
+/// Index 4: Fifth item in list.
+/// Index -1: Last item in list.
+/// Index -3: Third to last item in list.
+/// Index 1 Count 2: Second and third items in list.
+/// Index -3 Count 3: Last three items in list.
+/// Index -5 Count 4: Range between fifth to last item to second to last item inclusive.
+///
+///
+/// Nested expressions are supported by optional CTX context arguments.
 pub struct ListExpression {}
 
 const MODULE: i64 = 0;
 
+#[doc(hidden)]
 pub enum ListExpOp {
     APPEND = 1,
     AppendItems = 2,
@@ -37,6 +63,7 @@ pub enum ListExpOp {
     RemoveByValueRelRankRange = 40,
 }
 impl ListExpression {
+    /// Create expression that appends value to end of list.
     pub fn append(
         policy: ListPolicy,
         value: FilterExpression,
@@ -53,6 +80,7 @@ impl ListExpression {
         ListExpression::add_write(bin, ctx, args)
     }
 
+    /// Create expression that appends list items to end of list.
     pub fn append_items(
         policy: ListPolicy,
         list: FilterExpression,
@@ -69,6 +97,7 @@ impl ListExpression {
         ListExpression::add_write(bin, ctx, args)
     }
 
+    /// Create expression that inserts value to specified index of list.
     pub fn insert(
         policy: ListPolicy,
         index: FilterExpression,
@@ -86,6 +115,7 @@ impl ListExpression {
         ListExpression::add_write(bin, ctx, args)
     }
 
+    /// Create expression that inserts each input list item starting at specified index of list.
     pub fn insert_items(
         policy: ListPolicy,
         index: FilterExpression,
@@ -103,6 +133,8 @@ impl ListExpression {
         ListExpression::add_write(bin, ctx, args)
     }
 
+    /// Create expression that increments list[index] by value.
+    /// Value expression should resolve to a number.
     pub fn increment(
         policy: ListPolicy,
         index: FilterExpression,
@@ -121,6 +153,7 @@ impl ListExpression {
         ListExpression::add_write(bin, ctx, args)
     }
 
+    /// Create expression that sets item value at specified index in list.
     pub fn set(
         policy: ListPolicy,
         index: FilterExpression,
@@ -138,6 +171,7 @@ impl ListExpression {
         ListExpression::add_write(bin, ctx, args)
     }
 
+    /// Create expression that removes all items in list.
     pub fn clear(bin: FilterExpression, ctx: &[CdtContext]) -> FilterExpression {
         let args = vec![
             ExpressionArgument::Value(Value::from(ListExpOp::CLEAR as i64)),
@@ -146,6 +180,7 @@ impl ListExpression {
         ListExpression::add_write(bin, ctx, args)
     }
 
+    /// Create expression that sorts list according to sortFlags.
     pub fn sort(
         sort_flags: ListSortFlags,
         bin: FilterExpression,
@@ -159,6 +194,7 @@ impl ListExpression {
         ListExpression::add_write(bin, ctx, args)
     }
 
+    /// Create expression that removes list items identified by value.
     pub fn remove_by_value(
         value: FilterExpression,
         bin: FilterExpression,
@@ -173,6 +209,7 @@ impl ListExpression {
         ListExpression::add_write(bin, ctx, args)
     }
 
+    /// Create expression that removes list items identified by values.
     pub fn remove_by_value_list(
         values: FilterExpression,
         bin: FilterExpression,
@@ -187,6 +224,9 @@ impl ListExpression {
         ListExpression::add_write(bin, ctx, args)
     }
 
+    /// Create expression that removes list items identified by value range (valueBegin inclusive, valueEnd exclusive).
+    /// If valueBegin is null, the range is less than valueEnd. If valueEnd is null, the range is
+    /// greater than equal to valueBegin.
     pub fn remove_by_value_range(
         value_begin: Option<FilterExpression>,
         value_end: Option<FilterExpression>,
@@ -209,6 +249,18 @@ impl ListExpression {
         ListExpression::add_write(bin, ctx, args)
     }
 
+    /// Create expression that removes list items nearest to value and greater by relative rank.
+    ///
+    /// Examples for ordered list [0,4,5,9,11,15]:
+    ///
+    /// (value,rank) = [removed items]
+    /// (5,0) = [5,9,11,15]
+    /// (5,1) = [9,11,15]
+    /// (5,-1) = [4,5,9,11,15]
+    /// (3,0) = [4,5,9,11,15]
+    /// (3,3) = [11,15]
+    /// (3,-3) = [0,4,5,9,11,15]
+    ///
     pub fn remove_by_value_relative_rank_range(
         value: FilterExpression,
         rank: FilterExpression,
@@ -225,6 +277,18 @@ impl ListExpression {
         ListExpression::add_write(bin, ctx, args)
     }
 
+    /// Create expression that removes list items nearest to value and greater by relative rank with a count limit.
+    ///
+    /// Examples for ordered list [0,4,5,9,11,15]:
+    ///
+    /// (value,rank,count) = [removed items]
+    /// (5,0,2) = [5,9]
+    /// (5,1,1) = [9]
+    /// (5,-1,2) = [4,5]
+    /// (3,0,1) = [4]
+    /// (3,3,7) = [11,15]
+    /// (3,-3,2) = []
+    ///
     pub fn remove_by_value_relative_rank_range_count(
         value_begin: FilterExpression,
         value_end: FilterExpression,
@@ -243,6 +307,7 @@ impl ListExpression {
         ListExpression::add_write(bin, ctx, args)
     }
 
+    /// Create expression that removes list item identified by index.
     pub fn remove_by_index(
         index: FilterExpression,
         bin: FilterExpression,
@@ -257,6 +322,7 @@ impl ListExpression {
         ListExpression::add_write(bin, ctx, args)
     }
 
+    /// Create expression that removes list items starting at specified index to the end of list.
     pub fn remove_by_index_range(
         index: FilterExpression,
         bin: FilterExpression,
@@ -271,6 +337,7 @@ impl ListExpression {
         ListExpression::add_write(bin, ctx, args)
     }
 
+    /// Create expression that removes "count" list items starting at specified index.
     pub fn remove_by_index_range_count(
         index: FilterExpression,
         count: FilterExpression,
@@ -287,6 +354,7 @@ impl ListExpression {
         ListExpression::add_write(bin, ctx, args)
     }
 
+    /// Create expression that removes list item identified by rank.
     pub fn remove_by_rank(
         rank: FilterExpression,
         bin: FilterExpression,
@@ -301,6 +369,7 @@ impl ListExpression {
         ListExpression::add_write(bin, ctx, args)
     }
 
+    /// Create expression that removes list items starting at specified rank to the last ranked item.
     pub fn remove_by_rank_range(
         rank: FilterExpression,
         bin: FilterExpression,
@@ -315,6 +384,7 @@ impl ListExpression {
         ListExpression::add_write(bin, ctx, args)
     }
 
+    /// Create expression that removes "count" list items starting at specified rank.
     pub fn remove_by_rank_range_count(
         rank: FilterExpression,
         count: FilterExpression,
@@ -331,6 +401,14 @@ impl ListExpression {
         ListExpression::add_write(bin, ctx, args)
     }
 
+    /// Create expression that returns list size.
+    ///
+    /// ```
+    /// // List bin "a" size > 7
+    /// use aerospike::exp::Expression;
+    /// use aerospike::exp::list_exp::ListExpression;
+    /// Expression::gt(ListExpression::size(Expression::list_bin("a".to_string()), &[]), Expression::int_val(7));
+    /// ```
     pub fn size(bin: FilterExpression, ctx: &[CdtContext]) -> FilterExpression {
         let args = vec![
             ExpressionArgument::Value(Value::from(ListExpOp::SIZE as i64)),
@@ -339,6 +417,19 @@ impl ListExpression {
         ListExpression::add_read(bin, ExpType::INT, args)
     }
 
+    /// Create expression that selects list items identified by value and returns selected
+    /// data specified by returnType.
+    ///
+    /// ```
+    /// // List bin "a" contains at least one item == "abc"
+    /// use aerospike::exp::Expression;
+    /// use aerospike::exp::list_exp::ListExpression;
+    /// use aerospike::operations::lists::ListReturnType;
+    /// Expression::gt(
+    ///   ListExpression::get_by_value(ListReturnType::Count, Expression::string_val("abc".to_string()), Expression::list_bin("a".to_string()), &[]),
+    ///   Expression::int_val(0));
+    /// ```
+    ///
     pub fn get_by_value(
         return_type: ListReturnType,
         value: FilterExpression,
@@ -354,6 +445,16 @@ impl ListExpression {
         ListExpression::add_read(bin, ListExpression::get_value_type(return_type), args)
     }
 
+    /// Create expression that selects list items identified by value range and returns selected data
+    /// specified by returnType.
+    ///
+    /// ```
+    /// // List bin "a" items >= 10 && items < 20
+    /// use aerospike::exp::list_exp::ListExpression;
+    /// use aerospike::operations::lists::ListReturnType;
+    /// use aerospike::exp::Expression;
+    /// ListExpression::get_by_value_range(ListReturnType::Values, Some(Expression::int_val(10)), Some(Expression::int_val(20)), Expression::list_bin("a".to_string()), &[]);
+    /// ```
     pub fn get_by_value_range(
         return_type: ListReturnType,
         value_begin: Option<FilterExpression>,
@@ -377,6 +478,8 @@ impl ListExpression {
         ListExpression::add_read(bin, ListExpression::get_value_type(return_type), args)
     }
 
+    /// Create expression that selects list items identified by values and returns selected data
+    /// specified by returnType.
     pub fn get_by_value_list(
         return_type: ListReturnType,
         values: FilterExpression,
@@ -392,6 +495,19 @@ impl ListExpression {
         ListExpression::add_read(bin, ListExpression::get_value_type(return_type), args)
     }
 
+    /// Create expression that selects list items nearest to value and greater by relative rank
+    /// and returns selected data specified by returnType.
+    ///
+    /// Examples for ordered list [0,4,5,9,11,15]:
+    ///
+    /// (value,rank) = [selected items]
+    /// (5,0) = [5,9,11,15]
+    /// (5,1) = [9,11,15]
+    /// (5,-1) = [4,5,9,11,15]
+    /// (3,0) = [4,5,9,11,15]
+    /// (3,3) = [11,15]
+    /// (3,-3) = [0,4,5,9,11,15]
+    ///
     pub fn get_by_value_relative_rank_range(
         return_type: ListReturnType,
         value: FilterExpression,
@@ -409,6 +525,19 @@ impl ListExpression {
         ListExpression::add_read(bin, ListExpression::get_value_type(return_type), args)
     }
 
+    /// Create expression that selects list items nearest to value and greater by relative rank with a count limit
+    /// and returns selected data specified by returnType.
+    ///
+    /// Examples for ordered list [0,4,5,9,11,15]:
+    ///
+    /// (value,rank,count) = [selected items]
+    /// (5,0,2) = [5,9]
+    /// (5,1,1) = [9]
+    /// (5,-1,2) = [4,5]
+    /// (3,0,1) = [4]
+    /// (3,3,7) = [11,15]
+    /// (3,-3,2) = []
+    ///
     pub fn get_by_value_relative_rank_range_count(
         return_type: ListReturnType,
         value: FilterExpression,
@@ -428,6 +557,19 @@ impl ListExpression {
         ListExpression::add_read(bin, ListExpression::get_value_type(return_type), args)
     }
 
+    /// Create expression that selects list item identified by index and returns
+    /// selected data specified by returnType.
+    ///
+    /// ```
+    /// // a[3] == 5
+    /// use aerospike::exp::{Expression, ExpType};
+    /// use aerospike::exp::list_exp::ListExpression;
+    /// use aerospike::operations::lists::ListReturnType;
+    /// Expression::eq(
+    ///   ListExpression::get_by_index(ListReturnType::Values, ExpType::INT, Expression::int_val(3), Expression::list_bin("a".to_string()), &[]),
+    ///   Expression::int_val(5));
+    /// ```
+    ///
     pub fn get_by_index(
         return_type: ListReturnType,
         value_type: ExpType,
@@ -444,6 +586,8 @@ impl ListExpression {
         ListExpression::add_read(bin, value_type, args)
     }
 
+    /// Create expression that selects list items starting at specified index to the end of list
+    /// and returns selected data specified by returnType .
     pub fn get_by_index_range(
         return_type: ListReturnType,
         index: FilterExpression,
@@ -459,6 +603,8 @@ impl ListExpression {
         ListExpression::add_read(bin, ListExpression::get_value_type(return_type), args)
     }
 
+    /// Create expression that selects "count" list items starting at specified index
+    /// and returns selected data specified by returnType.
     pub fn get_by_index_range_count(
         return_type: ListReturnType,
         index: FilterExpression,
@@ -476,6 +622,16 @@ impl ListExpression {
         ListExpression::add_read(bin, ListExpression::get_value_type(return_type), args)
     }
 
+    /// Create expression that selects list item identified by rank and returns selected
+    /// data specified by returnType.
+    ///
+    /// ```
+    /// // Player with lowest score.
+    /// use aerospike::exp::list_exp::ListExpression;
+    /// use aerospike::operations::lists::ListReturnType;
+    /// use aerospike::exp::{ExpType, Expression};
+    /// ListExpression::get_by_rank(ListReturnType::Values, ExpType::STRING, Expression::int_val(0), Expression::list_bin("a".to_string()), &[]);
+    /// ```
     pub fn get_by_rank(
         return_type: ListReturnType,
         value_type: ExpType,
@@ -492,6 +648,8 @@ impl ListExpression {
         ListExpression::add_read(bin, value_type, args)
     }
 
+    /// Create expression that selects list items starting at specified rank to the last ranked item
+    /// and returns selected data specified by returnType.
     pub fn get_by_rank_range(
         return_type: ListReturnType,
         rank: FilterExpression,
@@ -507,6 +665,8 @@ impl ListExpression {
         ListExpression::add_read(bin, ListExpression::get_value_type(return_type), args)
     }
 
+    /// Create expression that selects "count" list items starting at specified rank and returns
+    /// selected data specified by returnType.
     pub fn get_by_rank_range_count(
         return_type: ListReturnType,
         rank: FilterExpression,
@@ -524,6 +684,7 @@ impl ListExpression {
         ListExpression::add_read(bin, ListExpression::get_value_type(return_type), args)
     }
 
+    #[doc(hidden)]
     pub fn add_read(
         bin: FilterExpression,
         return_type: ExpType,
@@ -540,20 +701,19 @@ impl ListExpression {
         }
     }
 
+    #[doc(hidden)]
     pub fn add_write(
         bin: FilterExpression,
         ctx: &[CdtContext],
         arguments: Vec<ExpressionArgument>,
     ) -> FilterExpression {
-        let mut return_type: ExpType;
+        let return_type: ExpType;
         if ctx.is_empty() {
             return_type = ExpType::LIST
+        } else if (ctx[0].id & 0x10) == 0 {
+            return_type = ExpType::MAP;
         } else {
-            if (ctx[0].id & 0x10) == 0 {
-                return_type = ExpType::MAP;
-            } else {
-                return_type = ExpType::LIST;
-            }
+            return_type = ExpType::LIST;
         }
 
         FilterExpression {
@@ -567,6 +727,7 @@ impl ListExpression {
         }
     }
 
+    #[doc(hidden)]
     pub fn get_value_type(return_type: ListReturnType) -> ExpType {
         if (return_type as u8 & !(ListReturnType::Inverted as u8)) == ListReturnType::Values as u8 {
             ExpType::LIST
