+++ src/policy/batch_policy.rs
@@ -13,9 +13,8 @@
 // License for the specific language governing permissions and limitations under
 // the License.
 
+use crate::exp::FilterExpression;
 use crate::policy::{BasePolicy, Concurrency, PolicyLike};
-use crate::exp::exp::FilterExpression;
-use crate::query::Filter;
 
 /// `BatchPolicy` encapsulates parameters for all batch operations.
 pub struct BatchPolicy {
@@ -47,7 +46,7 @@ pub struct BatchPolicy {
     pub send_set_name: bool,
 
     /// Optional Filter Expression
-    pub filter_expression: Option<FilterExpression>
+    pub filter_expression: Option<FilterExpression>,
 }
 
 impl BatchPolicy {
@@ -57,7 +56,7 @@ impl BatchPolicy {
     }
 
     /// Get the current Filter Expression
-    pub fn filter_expression(&self) -> &Option<FilterExpression> {
+    pub const fn filter_expression(&self) -> &Option<FilterExpression> {
         &self.filter_expression
     }
 }
@@ -69,7 +68,7 @@ impl Default for BatchPolicy {
             concurrency: Concurrency::Sequential,
             allow_inline: true,
             send_set_name: false,
-            filter_expression: None
+            filter_expression: None,
         }
     }
 }
