+++ src/policy/query_policy.rs
@@ -13,8 +13,8 @@
 // License for the specific language governing permissions and limitations under
 // the License.
 
+use crate::exp::FilterExpression;
 use crate::policy::{BasePolicy, PolicyLike};
-use crate::exp::exp::FilterExpression;
 
 /// `QueryPolicy` encapsulates parameters for query operations.
 #[derive(Debug, Clone)]
@@ -37,7 +37,7 @@ pub struct QueryPolicy {
     pub fail_on_cluster_change: bool,
 
     /// Optional Filter Expression
-    pub filter_expression: Option<FilterExpression>
+    pub filter_expression: Option<FilterExpression>,
 }
 
 impl QueryPolicy {
@@ -47,7 +47,7 @@ impl QueryPolicy {
     }
 
     /// Get the current Filter Expression
-    pub fn filter_expression(&self) -> &Option<FilterExpression> {
+    pub const fn filter_expression(&self) -> &Option<FilterExpression> {
         &self.filter_expression
     }
 }
@@ -59,7 +59,7 @@ impl Default for QueryPolicy {
             max_concurrent_nodes: 0,
             record_queue_size: 1024,
             fail_on_cluster_change: true,
-            filter_expression: None
+            filter_expression: None,
         }
     }
 }
